requirements.txt
bidict==0.22.1
click==8.1.3
Flask==2.2.3
Flask-SocketIO==5.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.8.0
Werkzeug==2.2.3
